+++ Sources/KeyboardKit/_Deprecated/KeyboardLayoutItem+Deprecated.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+@available(*, deprecated, renamed: "KeyboardLayoutItem.Row")
+public typealias KeyboardLayoutItemRow = KeyboardLayoutItem.Row
+
+@available(*, deprecated, renamed: "KeyboardLayoutItem.Rows")
+public typealias KeyboardLayoutItemRows = KeyboardLayoutItem.Rows